support V2 API V1 compatibility for query
The db, q, and precision parameters are url-encoded in the body of the
POST to /query. Token authentication by Authorization header is
supported.

Known bug: if you do a SELECT * FROM ... query then the server will
timeout after about 90 seconds and give a 503. Don't use the wildcard *
and it will be fine.